Create README.md
The Luhn formula is a widely used system for validating identification numbers.

Using the original number, double the value of every other digit. Then add the values of the individual digits together - if a doubled value now has two digits, add the digits individually). The identification number is valid if the sum is divisible by 10.